### 2024-01-10 #6199 @FlorianK13
By searching for "energy" and "data" in the [the list of people](https://reviewers.joss.theoj.org/lookup) that have already agreed to review, I found the following potential reviewers:

- [adamrjensen](https://github.com/adamrjensen)
- [nick-gorman](https://github.com/nick-gorman)
- [Himscipy](https://github.com/Himscipy)
- [sevberg](https://github.com/sevberg)
- [jordanperr](https://github.com/jordanperr)
