### 2024-01-17 a63f739
Change data type of Netbetreiberpruefung #482
