+++ open_mastr/utils/orm.py
@@ -473,7 +473,7 @@ class GasStorageExtended(ParentAllTables, Base):
     EinheitMastrNummer = Column(String, primary_key=True)
     DatumLetzteAktualisierung = Column(DateTime(timezone=True))
     LokationMastrNummer = Column(String)
-    NetzbetreiberpruefungStatus = Column(Boolean)
+    NetzbetreiberpruefungStatus = Column(String)
     NetzbetreiberpruefungDatum = Column(Date)
     AnlagenbetreiberMastrNummer = Column(String)
     Land = Column(String)
@@ -537,7 +537,7 @@ class GasProducer(ParentAllTables, Base):
     EinheitMastrNummer = Column(String, primary_key=True)
     DatumLetzteAktualisierung = Column(DateTime(timezone=True))
     LokationMastrNummer = Column(String)
-    NetzbetreiberpruefungStatus = Column(Boolean)
+    NetzbetreiberpruefungStatus = Column(String)
     NetzbetreiberpruefungDatum = Column(Date)
     AnlagenbetreiberMastrNummer = Column(String)
     Land = Column(String)
@@ -578,7 +578,7 @@ class GasConsumer(ParentAllTables, Base):
     EinheitMastrNummer = Column(String, primary_key=True)
     DatumLetzteAktualisierung = Column(DateTime(timezone=True))
     LokationMastrNummer = Column(String)
-    NetzbetreiberpruefungStatus = Column(Boolean)
+    NetzbetreiberpruefungStatus = Column(String)
     NetzbetreiberpruefungDatum = Column(Date)
     AnlagenbetreiberMastrNummer = Column(String)
     Land = Column(String)
@@ -620,7 +620,7 @@ class ElectricityConsumer(ParentAllTables, Base):
     EinheitMastrNummer = Column(String, primary_key=True)
     DatumLetzteAktualisierung = Column(DateTime(timezone=True))
     LokationMastrNummer = Column(String)
-    NetzbetreiberpruefungStatus = Column(Boolean)
+    NetzbetreiberpruefungStatus = Column(String)
     NetzbetreiberpruefungDatum = Column(Date)
     AnlagenbetreiberMastrNummer = Column(String)
     Land = Column(String)
