### 2024-02-02 #6199 @ppxasjsm
Hi @FlorianK13 I'll be your editor for this submission.  If you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). You can search [the list of people](https://reviewers.joss.theoj.org/lookup) that have already agreed to review and may be suitable for this submission. Also I noticed there is a missing DOI in the paper. Could you please try and add this and run the check references option with editorialbot to make sure no more references are missing? If you have any questions at any point please get in touch (ideally using @).
