### 2024-02-02 #6199 @FlorianK13
Hi @ppxasjsm, thanks for being the editor. Regarding potential reviewers, I created this list a few comments above in this issue. Additionally, I'll check the missing DOI soon and add it.

> By searching for "energy" and "data" in the [the list of people](https://reviewers.joss.theoj.org/lookup) that have already agreed to review, I found the following potential reviewers:
>
>     * [adamrjensen](https://github.com/adamrjensen)
>
>     * [nick-gorman](https://github.com/nick-gorman)
>
>     * [Himscipy](https://github.com/Himscipy)
>
>     * [sevberg](https://github.com/sevberg)
>
>     * [jordanperr](https://github.com/jordanperr)
