### 2024-02-03 #6199 @ppxasjsm
Hi @adamrjensen, would you be available to review this JOSS submission?
