### 2024-02-03 #6199 @ppxasjsm
Hi @nick-gorman, any chance you would be available to review this JOSS submission?
