### 2024-02-03 #6199 @AdamRJensen
> Hi @adamrjensen, would you be available to review this JOSS submission?

I just started as a JOSS editor so Ill spend my JOSS time in that role in the upcoming time
