### 2024-02-19 OpenEnergyPlatform/open-MaStR#487 @FlorianK13
Deprecate Mastr Mirror via SOAP API from open-mastr
# Discussion
**Users that are affected by this deprecation can leave their comments in this issue.**
______________________________________________
Should we think about adding a deprecation warning to the soap API features `.download` and `.mirror`? We could print the message that those features will not be maintained in the future. If people still use them, they should tell us in an Issue.

# Explanation
Both functions are used to download large amount of data via the SOAP API. Since BNetzA has a bulk download for quite some time now, I don't think that we need to maintain this feature anymore.

@chrwm What do you think?

# Todos

- [ ] State very prominently in documentation that these features are deprecated, maybe even in the sidebar
- [x] Delete tests for the deprecated features
