### 2024-02-24 38da6e2
remove webscraping for URL
The URL format has been changed and no longer includes the random string.
This means the web scraping with BeauitfulSoup is no longer necessary.
