### 2024-03-08 OpenEnergyPlatform/open-MaStR#487 @FlorianK13
Currently mastr mirror tests are causing problems, so we should also delete those tests and rather focus our (limited) time on fixing and maintaining other parts of the code.
