+++ .pre-commit-config.yaml
@@ -3,4 +3,4 @@ repos:
     rev: 22.6.0
     hooks:
       - id: black
-        language_version: python3.10
+        language_version: python3.11
