### 2024-03-08 e2f6d3f
Add deprecation warning for MaStRMirror class #487
Additionally, black did reformatting.
