### 2024-03-08 5ccfb10
Correct docstring for mkdocs #487
Not related to this feature, but had to be done anyway.
