### 2024-03-08 4d39594
Add OFFIS as partner to README
