### 2024-03-09 1b55915
consider possibility of URL not being found
When the generated URL is not found, the download will be retried
with the URL for the previous day. This could be applicable if the
download is attempted before the new file is published.

If the re-tried download also fails, the process is aborted completely.
