### 2024-04-02 8cb3fc6
Delete AnzeigeEinerStilllegung #496
