### 2024-04-02 99b71b7
Delete ArtDerStilllegung #496
