### 2024-04-02 ca98ae5
Delete DatumBeginnVorlaeufigenOderEndgueltigenStilllegung #496
