### 2024-04-02 533df9e
Rename ZugeordneteWirkleistungWechselrichter #496
