### 2024-04-02 28c5593
Add DatumEndgueltigeStilllegung #496
* from bulk download
