### 2024-04-02 0a3da33
Add DatumEndgueltigeStilllegung #496
* from bulk download
