### 2024-04-02 e0ec101
Add WebportalDesNetzbetreibers & RegisternummerPraefix #496
* from bulk download
