+++ CITATION.cff
@@ -28,7 +28,7 @@ authors:
 title: "open-MaStR"
 type: software
 license: AGPL-3.0
-version: 0.14.1
+version: 0.14.2
 doi: 
-date-released: 2024-01-17
+date-released: 2024-04-04
 url: "https://github.com/OpenEnergyPlatform/open-MaStR/"
