### 2024-04-09 b3418d8
Rename workflows #505
So that they can be distinguished when looking at the github actions.
