+++ .github/workflows/ci-production.yml
@@ -1,4 +1,4 @@
-name: CI
+name: CI - Pytest Production
 
 on:
   workflow_dispatch:
