+++ CITATION.cff
@@ -30,5 +30,5 @@ type: software
 license: AGPL-3.0
 version: 0.14.2
 doi: 
-date-released: 2024-04-04
+date-released: 2024-04-10
 url: "https://github.com/OpenEnergyPlatform/open-MaStR/"
