+++ CHANGELOG.md
@@ -6,6 +6,12 @@ For each version important additions, changes and removals are listed here.
 The format is inspired from [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and the versioning aims to respect [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+
+## [v0.XX.X] - 202X-XX-XX
+### Added
+### Changed
+### Removed
+
 ## [v0.14.2] Maintenance - 2024-04-10
 ### Changed
 - Fix and add URLs of example projects in readme [#481](https://github.com/OpenEnergyPlatform/open-MaStR/pull/481)
