### 2024-04-23 4fa7f75
Add new table to docs #509
