### 2024-04-23 OpenEnergyPlatform/open-MaStR#487 @FlorianK13
This is merged on develop and will be released soon.
