+++ pyproject.toml
@@ -1,6 +1,6 @@
 [build-system]
-requires = ["twine"]
-build-backend = "twine.build"
+requires = ["setuptools", "setuptools-scm"]
+build-backend = "setuptools.build_meta"
 
 [project]
 name = "open_mastr"
@@ -29,8 +29,8 @@ authors = [
 ]
 
 maintainers = [
-  {name = "Florian Kotthoff"}
-  {name = "Christoph Muschner", email = "datenzentrum@rl-institut.de"}
+  {name = "Florian Kotthoff"},
+  {name = "Christoph Muschner", email = "datenzentrum@rl-institut.de"},
   {name = "Ludwig Hülk", email = "datenzentrum@rl-institut.de"}
 ]
 description = "A package that provides an interface for downloading and processing the data of the Marktstammdatenregister (MaStR)"
@@ -79,15 +79,8 @@ open_mastr = [
   "soap_api/metadata/LICENSE",
 ]
 
+[tool.setuptools.packages.find]
+include = ["open_mastr", "open_mastr.soap_api", "open_mastr.soap_api.metadata", "open_mastr.utils", "open_mastr.utils.config", "open_mastr.xml_download"] # package names should match these glob patterns (["*"] by default)
+
 # from setup.py - not yet included in here
 # download_url="https://github.com/OpenEnergyPlatform/open-MaStR/archive""/refs/tags/v0.14.2.tar.gz",
-# setup(
-#    name="open_mastr",
-#    packages=[
-#        "open_mastr",
-#        "open_mastr.soap_api",
-#        "open_mastr.soap_api.metadata",
-#        "open_mastr.utils",
-#        "open_mastr.utils.config",
-#        "open_mastr.xml_download",
-#    ],
