+++ .github/workflows/ci-production.yml
@@ -28,6 +28,7 @@ jobs:
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
+        pip install build
     - name: create package
       run: python -m build --sdist
     - name: import open-mastr
