+++ .bumpversion.cfg
@@ -15,7 +15,7 @@ values =
 [bumpversion:part:build]
 first_value = 1
 
-[bumpversion:file:setup.py]
+[bumpversion:file:pyproject.toml]
 
 [bumpversion:file:./.github/workflows/ci-production.yml]
 
