+++ CITATION.cff
@@ -28,7 +28,7 @@ authors:
 title: "open-MaStR"
 type: software
 license: AGPL-3.0
-version: 0.14.2
+version: 0.14.3a1
 doi: 
 date-released: 2024-04-10
 url: "https://github.com/OpenEnergyPlatform/open-MaStR/"
