+++ CITATION.cff
@@ -30,5 +30,5 @@ type: software
 license: AGPL-3.0
 version: 0.14.3a1
 doi: 
-date-released: 2024-04-10
+date-released: 2024-04-24
 url: "https://github.com/OpenEnergyPlatform/open-MaStR/"
