### 2024-04-24 451a775
Add changes from release to Changelog #513
