+++ pyproject.toml
@@ -41,7 +41,7 @@ classifiers = [
   "Development Status :: 4 - Beta",
   "Intended Audience :: Science/Research",
   "Topic :: Scientific/Engineering",
-  Topic :: Scientific/Engineering :: GIS,
+  "Topic :: Scientific/Engineering :: GIS",
   "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
   "Programming Language :: Python :: 3",
   "Programming Language :: Python :: 3.9",
