+++ CITATION.cff
@@ -28,7 +28,7 @@ authors:
 title: "open-MaStR"
 type: software
 license: AGPL-3.0
-version: 0.14.3a1
+version: 0.14.3
 doi: 
 date-released: 2024-04-24
 url: "https://github.com/OpenEnergyPlatform/open-MaStR/"
