### 2024-05-10 692f2ee
Add offis logo to readme #490
