### 2024-05-10 d392e62
Delete third party links in README #490
I think they are placed too prominenty, especially since we have no
influence on whether the stuff at those links is working.
