### 2024-05-16 #6758 @editorialbot
[REVIEW]: open-mastr: A Python Package to Download and Process the German Energy Registry Marktstammdatenregister
**Submitting author:** @FlorianK13 (<a href="http://orcid.org/0000-0003-3666-6122">Florian Kotthoff</a>)
**Repository:** https://github.com/OpenEnergyPlatform/open-MaStR
**Branch with paper.md** (empty if default branch): feature-462-paper-joss
**Version:** v0.14.4
