### 2024-05-17 e0b418e
Update dois in paper #462
As well as author affiliations.
