+++ paper.bib
@@ -8,13 +8,14 @@
   year    = {2019},
 }
 
-@misc{Tepe2023_MaStR,
-      title={Improving Data Reliability in Germany's Energy System: A Validation of Unit Locations of the Marktstammdatenregister}, 
-      author={Deniz Tepe and Florian Kotthoff and Christoph Muschner and Esther Vogt and Ludwig Hülk},
-      year={2023},
+@misc{Kotthoff2024_monitoring,
+      title={Monitoring Germany's Core Energy System Dataset: A Data Quality Analysis of the Marktstammdatenregister}, 
+      author={Florian Kotthoff and Christoph Muschner and Deniz Tepe and Esther Vogt and Ludwig Hülk},
+      year={2024},
       eprint={2304.10581},
       archivePrefix={arXiv},
-      primaryClass={eess.SY}
+      primaryClass={eess.SY},
+      doi = {10.48550/arXiv.2304.10581}
 }
 
 @article{Holmgren2018_pvlib,
@@ -36,7 +37,8 @@
   volume={236},
   pages={401--409},
   year={2019},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.apenergy.2018.11.097}
 }
 
 @software{Hulk2022_OpenEnergyFamily,
