### 2024-05-17 #6758 @FlorianK13
> @FlorianK13 could you double check the missing DOIs in your references please? There seems to be at least one suggested that seems valid.

I have added the missing DOI and rerun the editorialbot commands to generate the DOI report and the new paper version.
