### 2024-05-23 #6758 @cpschau
@FlorianK13 the example for exporting data in the "Getting Started" section appears to contain a small mistake. The first line should likely be tables=["wind", "grids"] instead of tables=["wind", "grid"] (https://open-mastr.readthedocs.io/en/latest/getting_started/#exporting-data).

Additionally, while the documentation provides an extensive explanation of the data model, the "Getting Started" section could be enhanced with an example that lists all available tables in the database e.g. `df = pd.read_sql_query("SELECT name FROM sqlite_master WHERE type='table';", con=db.engine)`.
