### 2024-05-23 #6758 @FlorianK13
Thanks for your comment @cpschau
Could you copy this to a new issue in the [open-mastr](https://github.com/OpenEnergyPlatform/open-MaStR/issues) repository? Then I can refer to it when fixing the things that you mention.

If it helps, you can see one [example issue](https://github.com/canimus/cuallee/issues/239) that I have created for another JOSS Review.
