### 2024-05-23 3b142c7
Make HEADERS a constant #489
Co-authored-by: chrwm <54852694+chrwm@users.noreply.github.com>
