### 2024-05-24 4f1573a
Add command to get all tables #522
