### 2024-05-24 OpenEnergyPlatform/open-MaStR#524 @rafmudaf
Also, I think your docs site looks really nice. It has the right amount of information, highly readable 👌
