### 2024-05-26 5112df9
improves metadata from pyproject.toml
- supported python versions are implicit generated from requires-python
- license identifier makes license better visible
