### 2024-05-27 OpenEnergyPlatform/open-MaStR#524 @FlorianK13
Thanks @rafmudaf for this issue, I'll work through your points within this week.
