### 2024-05-28 eaef41b
Add installation instruction to docs #524
