### 2024-05-28 24d23d5
Delete outdated lines from CONTRIBUTING.md #524
