### 2024-05-28 9958c16
Repair formatting of lists in docs #524
