### 2024-05-28 069533b
Add license classifier #525
Based on this instruction:
https://packaging.python.org/en/latest/guides/writing-pyproject-toml/#license
