+++ README.rst
@@ -20,7 +20,7 @@ open-mastr
    * - Tests
      - |badge_ci|
    * - Publication
-     - |badge_pypi|
+     - |badge_pypi| |badge_joss|
    * - Data Publication
      - |badge_zenodo|
    * - Development
@@ -183,5 +183,8 @@ Data
 
 .. |Total PyPI downloads| image:: https://static.pepy.tech/badge/open-mastr
     :target: https://pepy.tech/project/open-mastr
+
+.. |badge_joss| image: https://joss.theoj.org/papers/dc0d33e7dc74f7233e15a7b6fe0c7a3e/status.svg
+    :target: https://joss.theoj.org/papers/dc0d33e7dc74f7233e15a7b6fe0c7a3e
 
 
