### 2024-05-28 OpenEnergyPlatform/open-MaStR#524 @FlorianK13
@rafmudaf I have added the changes. They will be merged to develop after the review of @chrwm in #527
