### 2024-05-29 4498de2
Add USERS.cff and update CONTRIBUTING.md
