### 2024-05-29 OpenEnergyPlatform/open-MaStR#524 @FlorianK13
The changes were merged in #527
