### 2024-05-29 #6758 @FlorianK13
First of all, thanks to @rafmudaf and @cpschau for reviewing `open-mastr` 👍

All changes from the two Reviews are now merged on our develop branch. We need to make a new release so that those changes actually appear on our doc page. Since every Release takes some time and effort, I first want to make sure that @cpschau and @rafmudaf have finished their review? Maybe you can indicate it with a thumbs up on this comment?

If the review is finished, we can start the release procedure for open-mastr.
