### 2024-06-03 3a6ff6e
Create issue_template_user_kudos.md
