+++ .github/ISSUE_TEMPLATE/issue_template_user_kudos.md
@@ -1,0 +1,23 @@
+---
+name: User Kudos
+about: Share appreciation for the project!
+title: Add new user to USERS.cff
+labels: user
+assignees: ''
+
+---
+
+Thanks for supporting the project and helping to improve its recognition by joining USERS.cff.
+It helps the project quite a bit!
+
+We will add you to the list of valued users.
+
+Please, insert your information between the double quotes below - fill out at minimum "affiliation" :purple_heart:
+
+family-names: ""
+given-names: ""
+alias: ""
+affiliation: ""
+orcid: ""
+
+Thank you!
