+++ USERS.cff
@@ -1,8 +1,8 @@
 cff-version: 1.2.0
 message: "If you use this software, please leave your name for reference.
 authors:
-  - family-names: ""
-    given-names: ""
-    alias: ""
-    affiliation: "Reiner Lemoine Institut"
-    orcid: ""
+  - family-names: 
+    given-names: 
+    alias: 
+    affiliation: Reiner Lemoine Institut
+    orcid:
