### 2024-06-04 OpenEnergyPlatform/open-MaStR#530 @FlorianK13
Release Version v0.14.4
Release Date: 2024-XX-XX
Integrating JOSS Review

- [x] Add versions to `changelog.md`
- [x] Fix and update package metadata in `pyproject.toml`
- [x] Update `citation.cff`
- [ ] Release on github
- [ ] Give Credit to JOSS Reviewers on Release notes
