### 2024-06-04 #6758 @FlorianK13
The release is started and tracked in https://github.com/OpenEnergyPlatform/open-MaStR/issues/530
