### 2024-06-05 #6758 @FlorianK13
> > @rafmudaf and @cpschau, it looks like you have both completed your review checklist. Have all your concerns been addressed? There is still an open issue for @FlorianK13 to address from @rafmudaf? Once you are happy with everything can you please let me know here?
>
> Sorry for the late reply! I was absent for the last week. Besides the minor improvements to the documentation already made by @FlorianK13, I did not find any additional issues with the tool. However, I plan to spend a few more hours this week testing the API download method.

Allright, then we pause the release for now and wait for your review @cpschau 👍
