### 2024-06-05 OpenEnergyPlatform/open-MaStR#532 @cpschau
[JOSS Review] Broken link to API registration and usage examples
The documentation (https://open-mastr.readthedocs.io/en/latest/advanced/#soap-api-download) references a guide to the API registration of MaStR (https://www.marktstammdatenregister.de/MaStRHilfe/files/regHilfen/201108_Handbuch%20f%C3%BCr%20Registrierungen%20durch%20Dienstleister.pdf). This link seems to be broken. Possibly it should be replaced with https://www.marktstammdatenregister.de/MaStRHilfe/files/regHilfen/Handbuch_fuer_Registrierungen_durch_Dienstleister.pdf ?

Another broken link was found in the documentation (https://open-mastr.readthedocs.io/en/latest/reference/advanced/#open_mastr.soap_api.download.MaStRAPI) which is supposed to point to the [pre-defined SOAP API queries](https://www.marktstammdatenregister.de/MaStRHilfe/files/webdienst/Funktionen_MaStR_Webdienste_V1.2.39.html). Probably it could be updated to https://www.marktstammdatenregister.de/MaStRHilfe/files/webdienst/Funktionen_MaStR_Webdienste_V24.1.128.html .

This issue is part of a https://github.com/openjournals/joss-reviews/issues/6684
