### 2024-06-05 #6758 @cpschau
> > > @rafmudaf and @cpschau, it looks like you have both completed your review checklist. Have all your concerns been addressed? There is still an open issue for @FlorianK13 to address from @rafmudaf? Once you are happy with everything can you please let me know here?
> >
> >
> > Sorry for the late reply! I was absent for the last week. Besides the minor improvements to the documentation already made by @FlorianK13, I did not find any additional issues with the tool. However, I plan to spend a few more hours this week testing the API download method.
>
> Allright, then we pause the release for now and wait for your review @cpschau 👍

I just had a little play with the API download method as well as the `MaStRAPI class using some of the functions referenced [here](https://www.marktstammdatenregister.de/MaStRHilfe/files/webdienst/Funktionen_MaStR_Webdienste_V24.1.128.html). Everything seems to be functional. I only faced issues once I used the deprecated class `open_mastr.soap_api.download.MaStRDownload`, which is addressed in https://github.com/OpenEnergyPlatform/open-MaStR/issues/487.

Once the broken links in the documentation are fixed (https://github.com/OpenEnergyPlatform/open-MaStR/issues/532), I will be happy, and you can proceed with the new release.
