### 2024-06-06 6b61622
If applied, this commit will close issue #533.
If applied, this commit will close issue #533 by adding the missing call to gen_url.
