### 2024-06-06 07aac98
Run black #533
Additionally deleted two comments that I think are not necessary
to understand the code
