### 2024-06-06 OpenEnergyPlatform/open-MaStR#532 @FlorianK13
This was merged in #536
