+++ CITATION.cff
@@ -30,5 +30,5 @@ type: software
 license: AGPL-3.0
 version: 0.14.4
 doi: 
-date-released: 2024-04-24
+date-released: 2024-06-07
 url: "https://github.com/OpenEnergyPlatform/open-MaStR/"
