### 2024-06-07 4d54fd5
Update RELEASE_PROCEDURE #530
