### 2024-06-07 99f3a3e
Add new user to USERS.cff
