+++ USERS.cff
@@ -6,3 +6,8 @@ authors:
     alias: 
     affiliation: Reiner Lemoine Institut
     orcid:
+  - family-names: Kotthoff
+    given-names: Florian
+    alias: @FlorianK13
+    affiliation: OFFIS
+    orcid: https://orcid.org/0000-0003-3666-6122
