### 2024-06-07 6cce347
Fix formatting issue_template_user_kudos.md
