### 2024-06-07 7d615c0
Update issue_template_user_kudos.md
