### 2024-06-07 #6758 @FlorianK13
@ppxasjsm The release tag is [v0.14.4](https://github.com/OpenEnergyPlatform/open-MaStR/releases/tag/v0.14.4), on zenodo it should appear as https://doi.org/10.5281/zenodo.11516669 as soon as it was reviewed by the zenodo Open Energy Family Group. This Review will hopefully take place this evening.
