### 2024-06-07 d1e460b
Add e.V. to affiliation in USERS.cff #541
