+++ USERS.cff
@@ -9,5 +9,5 @@ authors:
   - family-names: Kotthoff
     given-names: Florian
     alias: @FlorianK13
-    affiliation: OFFIS
+    affiliation: OFFIS e.V.
     orcid: https://orcid.org/0000-0003-3666-6122
