### 2024-06-17 #6758 @FlorianK13
@ppxasjsm Are the release on github and zenodo everything you need from us, or is there anything else?
