### 2024-08-27 OpenEnergyPlatform/open-MaStR#487 @nesnoj
Leave open for user feedback
